requirements.txt
discord.py[voice]>=2.4.0
aiohttp>=3.9.0
PyNaCl>=1.5.0
async-timeout>=4.0.0
yt-dlp==2026.3.17